Remove BulletinsCommon

+++ app/controllers/web/bulletins_controller.rb
@@ -2,27 +2,28 @@
 
 module Web
   class BulletinsController < Web::ApplicationController
-    include BulletinsCommon
-
-    after_action :check_policy, only: %i[show new create edit update destroy]
-    before_action :set_bulletin, only: %i[show edit update destroy]
+    before_action :set_bulletin, only: %i[show edit update destroy archive to_moderate]
     before_action :set_categories, only: %i[new edit]
 
     def index
-      index_prepare
+      @q = Bulletin.published.includes(:user, :category).ransack(params[:q])
+      @bulletins = @q.result.page(params[:page]).order(created_at: :desc)
     end
 
     def show; end
 
     def new
       @bulletin = Bulletin.new
+      authorize @bulletin
     end
 
-    def edit; end
+    def edit
+      authorize @bulletin
+    end
 
     def create
       @bulletin = Bulletin.new({ user_id: current_user.id }.merge(bulletin_params))
-
+      authorize @bulletin
       if @bulletin.save
         redirect_to profile_path, notice: t('.success')
       else
@@ -32,6 +33,7 @@ module Web
     end
 
     def update
+      authorize @bulletin
       if @bulletin.update(bulletin_params)
         redirect_to profile_path, notice: t('.success')
       else
@@ -41,6 +43,7 @@ module Web
     end
 
     def destroy
+      authorize @bulletin
       if @bulletin.destroy
         redirect_to @bulletin, notice: t('.success')
       else
@@ -49,10 +52,38 @@ module Web
       end
     end
 
+    def archive
+      authorize @bulletin
+      if @bulletin.may_archive? && @bulletin.aasm.fire!(:archive)
+        redirect_to back_path, notice: t('.success')
+      else
+        flash[:error] = t('.error')
+        redirect_to back_path, status: :unprocessable_entity
+      end
+    end
+
+    def to_moderate
+      authorize @bulletin
+      if @bulletin.may_to_moderate? && @bulletin.aasm.fire!(:to_moderate)
+        redirect_to back_path, notice: t('.success')
+      else
+        flash[:error] = t('.error')
+        redirect_to back_path, status: :unprocessable_entity
+      end
+    end
+
     private
+
+    def set_bulletin
+      @bulletin = Bulletin.includes(:category).find(params[:id])
+    end
 
     def set_categories
       @categories = Category.where.not(name: nil)
+    end
+
+    def bulletin_params
+      params.require(:bulletin).permit(%i[category_id description image title])
     end
   end
 end
